some initial bug fixes

+++ src/constant/communication/general.py
@@ -4,6 +4,6 @@ _config = load_config()["communication"]
 
 BUFFER_SIZE: int = int(_config["buffer_size"])
 
-TIMEOUT_RECEIVE: float = float(_config["timeout_receive"])
-TIMEOUT_RESPONSE: float = float(_config["timeout_response"])
-TIMEOUT_REPLICATION: float = float(_config["timeout_replication"])
+TIMEOUT_RECEIVE: int = int(_config["timeout_receive"])
+TIMEOUT_RESPONSE: int = int(_config["timeout_response"])
+TIMEOUT_REPLICATION: int = int(_config["timeout_replication"])
